- added gradient descent optimizer to allocate weights in conflicting situations of transitions that have (mainly non-free choice nets, or preselection policy)
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/StochasticPetriNets/Trunk@9103 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/models/graphbased/directed/petrinet/elements/TimedTransition.java
@@ -210,7 +210,7 @@ public class TimedTransition extends Transition{
 		distributionParameters = parameters;
 	}
 
-	public void setWeight(int weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
